// src/model/movie.tsx
export interface IMovie {
    created?: string;
    director?: string;
    edited?: string;
    episode_id?: string;
    price?: number;
    discount?: number;
    producer?: string;
    release_date?: string;
    title?: string;
    url?: any;
    opening_crawl?: string;
    characters?: string[];
};